backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.10.0
python-multipart==0.0.6
pypdf==4.0.0
python-docx==1.1.0
python-dotenv==1.0.0
